Abandoned references to objects and objectlists
Instead of trying to implement the Insert, Update and Delete in the
super classes of the table object super type, I now am leaning towards
generating the code in the generated classes instead, så that each
generated type will be responsible for storing and retrieving it's
members.

The code is now in a state where the super objects are near done, but
they need a final walkthrough, perhaps two.